Dharma-Platform-Live-Dashboard-master/services/stream-processing-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kafka-python==2.0.2
structlog==23.2.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
aiohttp==3.9.1
asyncio-mqtt==0.16.1
redis==5.0.1
pymongo==4.6.0
psycopg2-binary==2.9.9
elasticsearch==8.11.0